changed timestamps from float128 to float64 by saving it in ns

+++ espargos/backlog.py
@@ -28,7 +28,7 @@ class CSIBacklog(object):
         self.storage_ht40 = np.zeros((size,) + self.pool.get_shape() + ((csi.csi_buf_t.htltf_lower.size + csi.HT40_GAP_SUBCARRIERS * 2 + csi.csi_buf_t.htltf_higher.size) // 2,), dtype = np.complex64)
         self.storage_lltf = np.zeros((size,) + self.pool.get_shape() + (csi.csi_buf_t.lltf.size // 2,), dtype = np.complex64)
 
-        self.storage_timestamps = np.zeros((size,) + self.pool.get_shape(), dtype = np.float128)
+        self.storage_timestamps = np.zeros((size,) + self.pool.get_shape(), dtype = np.float64)
         self.storage_rssi = np.zeros((size,) + self.pool.get_shape(), dtype = np.float32)
         self.storage_macs = np.zeros((size, 6), dtype = np.uint8)
         self.head = 0
